int -> size_t

+++ lib/service/servicemp3.cpp
@@ -496,7 +496,7 @@ RESULT eServiceMP3::info(ePtr<iServiceInformation>&i)
 RESULT eServiceMP3::getName(std::string &name)
 {
 	name = m_filename;
-	int n = name.rfind('/');
+	size_t n = name.rfind('/');
 	if (n != std::string::npos)
 		name = name.substr(n + 1);
 	return 0;
